Add libbpf_set_max_verifier_log_output()
The global function is used to limit the BPF verifier log output. It
ensures that max last characters will only be printed.

The change hasn't been upstreamed yet. Once it has been, the commit will
be replaced (this is solely to avoid generating large sysdumps in the CI
when testing the unreleased v1.11 cilium).

Signed-off-by: Martynas Pumputis <m@lambda.lt>